Don't print headers in PDF
By default, Chrome inserts header and footer into the printed PDF. They contain the current time and URL.

When our customers use PDF printing for server-side rendering, they want to hide this information. In real life, they will contain the server time and the path to a temporary page, which are irrelevant to their end users.